Move navigation components to their own folder in lib-react-interbit
