+++ packages/lib-react-interbit/src/components/Navigation/Navigation.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types'
 import { Navbar, Nav, NavItem } from 'react-bootstrap'
 import { LinkContainer } from 'react-router-bootstrap'
 
-import NavWrap from '../NavWrap'
+import NavWrap from './NavWrap'
 import LinkWrapper from '../UIKit/LinkWrapper'
 import Logo from '../UIKit/Logo'
