+++ packages/lib-react-interbit/src/components/Navigation/NavigationWrapper.js
@@ -1,21 +1,19 @@
-// Header component for app-interbit.io
-// TODO: refactor and combine with other apps' header nav
-
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import { Grid, Row, Col, Navbar, NavItem, Nav } from 'react-bootstrap'
 import { LinkContainer } from 'react-router-bootstrap'
 
 import NavWrap from './NavWrap'
-import LinkWrapper from './UIKit/LinkWrapper'
-import Logo from './UIKit/Logo'
-import IBIcon from './UIKit/IBIcon'
+import Footer from './Footer'
+import Logo from '../UIKit/Logo'
+import IBIcon from '../UIKit/IBIcon'
 
-export default class HeaderNav extends Component {
+export default class NavigationWrapper extends Component {
   static propTypes = {
     account: PropTypes.shape({
+      userIcon: PropTypes.oneOfType([PropTypes.string, PropTypes.element]),
       to: PropTypes.string,
-      text: PropTypes.string
+      logout: PropTypes.string
     }),
     headerNavItems: PropTypes.arrayOf(
       PropTypes.shape({
@@ -25,15 +23,26 @@ export default class HeaderNav extends Component {
         isHidden: PropTypes.bool
       })
     ),
+    footerNavItems: PropTypes.arrayOf(PropTypes.shape({})), // TODO: Define the shape
+    footerBottomLinks: PropTypes.arrayOf(
+      PropTypes.shape({
+        to: PropTypes.string,
+        text: PropTypes.string
+      })
+    ),
+    container: PropTypes.element.isRequired,
     className: PropTypes.string
   }
 
   static defaultProps = {
     account: {
-      to: '#',
-      text: ''
+      userIcon: '',
+      accountTo: '#',
+      logout: '#'
     },
     headerNavItems: [],
+    footerNavItems: [],
+    footerBottomLinks: [],
     className: ''
   }
 
@@ -56,7 +65,14 @@ export default class HeaderNav extends Component {
       )
     }
 
-    const { headerNavItems, account, className } = this.props
+    const {
+      container,
+      headerNavItems,
+      footerNavItems,
+      footerBottomLinks,
+      account,
+      className
+    } = this.props
 
     return (
       <div className={`ibweb ${className}`}>
@@ -74,19 +90,24 @@ export default class HeaderNav extends Component {
                     </Navbar.Brand>
                   </Navbar.Header>
                   <Nav className="nav-main-menu">
-                    {headerNavItems.length &&
-                      headerNavItems.map(
-                        navItem => !navItem.isHidden && renderNavItem(navItem)
-                      )}
+                    {headerNavItems.map(
+                      navItem => !navItem.isHidden && renderNavItem(navItem)
+                    )}
                   </Nav>
                 </Navbar>
               </Col>
               <Col lg={2} md={3} sm={12} className="accounts-link">
-                <LinkWrapper to={account.to}>{account.text}</LinkWrapper>
+                {account.userIcon}
               </Col>
             </Row>
           </Grid>
         </div>
+        <Grid className="ibweb-container">
+          {container}
+          {footerNavItems && (
+            <Footer sections={footerNavItems} bottomLinks={footerBottomLinks} />
+          )}
+        </Grid>
       </div>
     )
   }
