+++ packages/lib-react-interbit/src/components/Navigation/SideBarNav.js
