--- packages/lib-react-interbit/src/components/UIKit/SidebarNav.md
@@ -1,4 +1,0 @@
-SidebarNav example:
-```js
-
-```
